+++ lib/src/ui/page/photo_main_page.dart
@@ -65,6 +65,8 @@ class _PhotoMainPageState extends State<PhotoMainPage>
   String get currentGalleryName {
     if (currentPath?.isAll == true) {
       return i18nProvider.getAllGalleryText(options);
+    } else if (currentPath == null) {
+      return i18nProvider.getNoSelectedText(options);
     }
     return currentPath?.name ?? "Select Folder";
   }
@@ -217,7 +219,7 @@ class _PhotoMainPageState extends State<PhotoMainPage>
         pathList = await PhotoManager.getAssetPathList(type: RequestType.image);
         break;
       case PickType.onlyVideo:
-        pathList = await PhotoManager.getAssetPathList(type: RequestType.image);
+        pathList = await PhotoManager.getAssetPathList(type: RequestType.video);
         break;
       default:
         pathList = await PhotoManager.getAssetPathList(